Initial preview control
