+++ Source/Pawnmorphs/Esoteria/Pawnmorph.csproj
@@ -876,8 +876,8 @@
     <Compile Include="User Interface\MutationData.cs" />
     <Compile Include="User Interface\Preview\DrawOptions.cs" />
     <Compile Include="User Interface\Preview\PawnPreview.cs" />
-    <Compile Include="User Interface\Preview\PreviewBase.cs" />
-    <Compile Include="User Interface\Preview\RacePreview.cs" />
+    <Compile Include="User Interface\Preview\Preview.cs" />
+    <Compile Include="User Interface\Preview\ThingPreview.cs" />
     <Compile Include="User Interface\Settings\Dialog_OptionalPatches.cs" />
     <Compile Include="User Interface\Settings\Dialog_RaceReplacements.cs" />
     <Compile Include="User Interface\Settings\Dialog_VisibleRaceSelection.cs" />
